test update to avoid failed to delete if there is not path

+++ Tests/CatalogItems/Rest/Out-RsRestCatalogItem.Tests.ps1
@@ -44,7 +44,10 @@ Describe "Out-RsRestCatalogItem" {
 
     AfterEach {
         Remove-RsCatalogItem -ReportServerUri $reportServerUri -RsFolder $rsFolderPath -Confirm:$false
-        Remove-Item -Path $localFolderPath -Recurse
+        if ($localFolderPath -ne "")
+        {
+            Remove-Item -Path $localFolderPath -Recurse
+        }
     }
 
     Context "ReportPortalUri parameter" {
